potion-analytics/potion-analytics-thicc/requirements.txt

numpy >= 1.20.3
scipy >= 1.7.1
pandas >= 1.3.2
scikit-learn >= 0.24.2
requests >= 2.27.1
requests-toolbelt >= 0.9.1
plotly >= 5.1.0
vaex >= 4.5.0
vaex-hdf5 >= 0.10.0
h5py >= 3.4.0
vaex-core >= 4.5.1
kaleido >= 0.2.1
Pillow >= 8.3.1
streamlit >= 0.89.0
yfinance >= 0.1.63
pdoc3 >= 0.8.1
pycoingecko >= 2.2.0
gql >= 2.0.0
graphql-core >= 2.3.2
streamlit-aggrid >= 0.2.2.post4
git+git://github.com/gregversteeg/NPEET@d562b77#egg=npeet